Implemented public transport cost evaluator

+++ SPDB/project/data-access/src/main/java/pl/edu/pw/elka/spdb/evaluators/PublicTransportCostEvaluator.java
@@ -39,7 +39,7 @@ public class PublicTransportCostEvaluator implements CostEvaluator<Double> {
             }
         }
 
-        Double cost = (Double) relationship.getProperty("duration");
+        Double cost = ((Long) relationship.getProperty("duration")).doubleValue();
 
         if (needChange) {
             cost += timeNeededToChange.getSeconds();
